StreetCars

Documentation for StreetCars.

StreetCars.StreetCars

StreetCars.RouteGrid

StreetCars.add_street_to_seen

StreetCars.change_duration

StreetCars.directed_random_walk

StreetCars.directed_random_walk

StreetCars.generic_walk

StreetCars.get_seen_and_unseen_canditates

StreetCars.upper_bound

StreetCars.StreetCars — Module

StreetCars

Package designed to interact with the data of the 2014 Google Hash Code Project

StreetCars.RouteGrid — Type

RouteGrid

Store a city along with additional instance parameters.

Fields

city::City: the city

seen_streets::Set{Int}: the set of streets that have been seen

StreetCars.add_street_to_seen — Method

add_street_to_seen(rg, street)

Add a street to the set of seen streets in a city.

StreetCars.change_duration — Method

change_duration(rg, total_duration)

Create a new RouteGrid with a different total_duration and everything else equal.

StreetCars.directed_random_walk — Method

directed_random_walk(rng, city)
directed_random_walk(city)

Create a solution from a City by letting each car follow a random walk from its starting point. Bias is given to streets that have not been seen yet.

StreetCars.directed_random_walk — Method

random_walk(rg)

Create a solution from a City by letting each car follow a random walk from its starting point.

StreetCars.generic_walk — Method

generic_walk(rg, search_function)

Create a solution from a City by letting each car follow a walk from its starting point given a search function.

StreetCars.get_seen_and_unseen_canditates — Method

get_seen_and_unseen_canditates(rg, current_junction, duration)

Get the seen and unseen candidates for the next street to take in a city.

StreetCars.upper_bound — Method

upper_bound(city)

Estimates the upper bound on the total distance that can be covered by 8 cars for a given city. Note that the city includes the total duration, the number of cars, the starting junction, and the streets.

The method involves sorting the city's streets by their distance to duration ratio and then greedily choosing the streets that have the greatest ratio until the total duration of the city is reached. Note that since we have 8 cars, we can consider the total duration to be 8 times the duration of the city.