API Reference

StreetCars.StreetCars — Module

StreetCars

Package designed to interact with the data of the 2014 Google Hash Code Project

StreetCars.RouteGrid — Type

RouteGrid

A immutable custom type that stores both a HashCode2014 City along with a set of streets that have been seen by street cars that are exploring that City.

Fields

city::City: the HashCode2014 City instance that represents the current problem

seen_streets::Set{Int}: the set of streets that have been seen

StreetCars.add_street_to_seen — Method

add_street_to_seen(rg, street)

Add the given street to the set of seen streets in the given RouteGrid.

StreetCars.change_duration — Method

change_duration(rg, total_duration)

Create and returns a new RouteGrid with the specified total_duration. All other fields should be the same as the given RouteGrid rg.

StreetCars.directed_random_walk — Method

directed_random_walk(rng, city)
directed_random_walk(city)

Create a solution from a City by letting each car follow a random walk from its starting point. Bias is given to streets that have not been seen yet.

StreetCars.directed_random_walk — Method

random_walk(rg)

Create a solution from a RouteGrid by letting each car follow a random walk from its starting point. Calls generic_walk with a search function that chooses a random candidate.

StreetCars.generic_walk — Method

generic_walk(rg, search_function)

Create a solution from a RouteGrid by letting each car follow a walk from its starting point given a search function. The search function should take in an array of candidates and return a tuple of the index of the chosen candidate and the candidate itself.

The walk algorithm works as follows:

Initialize the itinerary of each car to be a list containing only the starting junction.

While there are still streets that can be explored for a current car: a. Get the current junction of the car. b. Get the candidates and unseen candidates for the next street to take. c. If there are no unseen candidates, then use the seen candidates. If there are no seen candidates, then break out of the loop. d. Use the search function to choose the next street to take. e. Add the chosen street to the itinerary of the car. f. Add the chosen street to the set of seen streets.

StreetCars.get_seen_and_unseen_canditates — Method

get_seen_and_unseen_canditates(rg, current_junction, duration)

Given a RouteGrid rg, a current junction current_junction, and a duration duration, return a tuple of two arrays of tuples. The first array contains all the candidates for the next street to take in a city. A candidate is any street that starts at the current junction and whose duration does not exceed the total duration of the city. The second returned array contains all of the candidates streets that have not been seen in rg.

StreetCars.upper_bound — Method

upper_bound(city)

Estimates the upper bound on the total distance that can be covered by 8 cars for a given city. Note that the city includes the total duration, the number of cars, the starting junction, and the streets.

The method involves sorting the city's streets by their distance to duration ratio and then greedily choosing the streets that have the greatest ratio until the total duration of the city is reached. Note that since we have 8 cars, we can consider the total duration to be 8 times the duration of the city.

HashCode2014.HashCode2014 — Module

HashCode2014

Lightweight package designed to interact with the data of the 2014 Google Hash Code.

Challenge description: https://storage.googleapis.com/coding-competitions.appspot.com/HC/2014/hashcode2014_final_task.pdf

Challenge data: https://storage.googleapis.com/coding-competitions.appspot.com/HC/2014/paris_54000.txt

Made for MIT course C25: Julia: Solving Real-World Problems with Computation

HashCode2014.City — Type

City

Store a city made of Junctions and Streets, along with additional instance parameters.

Fields

total_duration::Int: total time allotted for the car itineraries (in seconds)

nb_cars::Int: number of cars in the fleet

starting_junction::Int: junction at which all the cars are located initially

junctions::Vector{Junction}: list of junctions

streets::Vector{Street}: list of streets

HashCode2014.Junction — Type

Junction

Store a city junction.

Fields

latitude::Float64: latitude (in decimal degrees)

longitude::Float64: longitude (in decimal degrees)

HashCode2014.Solution — Type

Solution

Store a set of itineraries, one for each car.

Fields

itineraries::Vector{Vector{Int}}: each itinerary is a vector of junction indices

HashCode2014.Street — Type

Street

Store an edge between two Junctions.

Fields

endpointA::Int: index of the first junction

endpointB::Int: index of the second junction

bidirectional::Bool: whether B -> A is allowed

duration::Int: time cost of traversing the street (in seconds)

distance::Int: length of the street (in meters)

HashCode2014.get_street_end — Method

get_street_end(i, street)

Retrieve the arrival endpoint of street when it starts at junction i.

HashCode2014.is_feasible — Method

is_feasible(solution, city[; verbose=false])

Check if solution satisfies the constraints for the instance defined by city. The following criteria are considered (taken from the problem statement):

the number of itineraries has to match the number of cars of city

the first junction of each itinerary has to be the starting junction of city

for each consecutive pair of junctions on an itinerary, a street connecting these junctions has to exist in city (if the street is one directional, it has to be traversed in the correct direction)

the duration of each itinerary has to be lower or equal to the total duration of city

HashCode2014.is_street — Method

is_street(i, j, street)

Check if the trip from junction i to junction j corresponds to a valid direction of street.

HashCode2014.is_street_start — Method

is_street_start(i, street)

Check if junction i corresponds to a valid starting point of street.

HashCode2014.plot_streets — Function

plot_streets(city, solution=nothing; path=nothing)

Plot a City and an optional Solution using the Python library folium, save the result as an HTML file at path.

HashCode2014.read_city — Function

read_city(path)

Read and parse a City from a file located at path.

The default path is an artifact containing the official challenge data.

HashCode2014.read_solution — Method

read_solution(solution, path)

Read and parse a Solution from a file located at path.

HashCode2014.total_distance — Method

total_distance(solution, city)

Compute the total distance of all itineraries in solution based on the street data from city. Streets visited several times are only counted once.

HashCode2014.write_city — Method

write_city(city, path)

Write a City to a file located at path.

HashCode2014.write_solution — Method

write_solution(solution, path)

Write a Solution to a file located at path.

HashCode2014.HashCode2014

StreetCars.StreetCars

HashCode2014.City

HashCode2014.Junction

HashCode2014.Solution

HashCode2014.Street

StreetCars.RouteGrid

HashCode2014.get_street_end

HashCode2014.is_feasible

HashCode2014.is_street

HashCode2014.is_street_start

HashCode2014.plot_streets

HashCode2014.read_city

HashCode2014.read_solution

HashCode2014.total_distance

HashCode2014.write_city

HashCode2014.write_solution

StreetCars.add_street_to_seen

StreetCars.change_duration

StreetCars.directed_random_walk

StreetCars.directed_random_walk

StreetCars.generic_walk

StreetCars.get_seen_and_unseen_canditates

StreetCars.upper_bound